+++ R/ellipse.axes.R
@@ -33,6 +33,7 @@
 #' @param label.pos Positions of text labels relative to the ends of the axes used in \code{\link[graphics]{text}} for
 #'          the four possible \code{label.ends}. 1, 2, 3, 4 represent below, to the left, above and to the right.
 #'          The default, \code{c(2, 4, 1, 3)}, positions the labels outside the axes.
+#' @param type Character. Draw \code{"lines"} or \code{"arrows"}.
 #' @param \dots Other arguments passed to \code{\link[graphics]{lines}} and \code{\link[graphics]{text}}.
 #'
 #' @return Invisibly returns a 4 x 2 matrix containing the end points of the axes in pairs (min, max) by rows.
@@ -49,25 +50,44 @@
 #' plot(iris[,1:2], asp=1)
 #' car::ellipse(mu, cov, radius = radius)
 #' res <- ellipse.axes(cov, center=mu, level = 0.68,
-#'             labels = TRUE)
+#'                     labels = TRUE)
 #' res
+#'
+#' # try some options 
+#' plot(iris[,1:2], asp=1)
+#' car::ellipse(mu, cov, radius = radius)
+#' abline(h=mu[2], v=mu[1], col = "grey")
+#' ellipse.axes(cov, centre=mu, level = 0.68,
+#'              labels = "Dim", label.ends = 1:4,
+#'              lwd = 2, lty = 2, col = "red",
+#'              cex = 1.5)
+#'              
+#' # draw arrows rather than lines            
+#' plot(iris[,1:2], asp=1)
+#' car::ellipse(mu, cov, radius = radius)
+#' ellipse.axes(cov, center=mu, level = 0.68,
+#'              type = "arrows")
 
 
 ellipse.axes <- function(
     x, 
-    centre = c(0, 0), 
-    center = center,
-    scale = c(1, 1), 
+    centre = c(0, 0),
+    center = centre,
+    scale, 
     which = 1:2, 
     level = 0.95,
     radius = sqrt(qchisq(level, 2)), 
     labels = TRUE, 
     label.ends = c(2, 4),
     label.pos = c(2, 4, 1, 3),
+    type = c("lines", "arrows"),
     ...) 
 {
   stopifnot(is.matrix(x)) 
   stopifnot(dim(x)[1] ==  dim(x)[2])  # square matrix?
+  if (!missing(centre) && !missing(center)) {
+    warning("Specify 'centre' or 'center', not both. The value from 'center' will be used.")
+  }
 
   if(length(which) != 2)  stop("`which` must be a vector of length 2, not", which)
   cov <- x[which, which]
@@ -81,13 +101,12 @@ ellipse.axes <- function(
        0, 1), nrow = 4, ncol = 2, byrow = TRUE)
   rownames(axes)<- apply(expand.grid(c("min","max"), c("X","Y"))[,2:1],
                          1, paste, collapse="")
-
   # transform to PC axes
   result <- axes %*% sqrt(diag(eig$values)) %*% t(eig$vectors)
   # scale
   result <- result %*% diag(rep(radius, 2))
-  # center 
-  result <- sweep(result, 2L, centre, FUN="+")
+  # center
+  result <- sweep(result, 2L, center, FUN="+")
   colnames(result) <- colnames(cov)
   
   # fix min/max
@@ -97,38 +116,47 @@ ellipse.axes <- function(
   # convert to something that can be used by lines()
   #result <- rbind(result[1:2,], NA, result[3:4,])
 
-  # draw the lines, inserting NA between each pair
-  lines(rbind(result[1:2,], NA, result[3:4,]), ...)
+  # draw the lines (inserting NA between each pair) or arrows
+  type <- match.arg(type)
+  if (type == "lines") {
+    lines(rbind(result[1:2,], NA, result[3:4,]), ...)
+  }
+  else {
+    arrows(center[1], center[2], 
+           x1 = result[,1], y1 = result[,2], 
+           angle = 12, length = .15, ...)
+  }
 
   if (!missing(labels)) {
-    if (is.logical(labels) & labels) labels <- paste("PC", 1:2, sep="")
+    if (is.logical(labels) & isTRUE(labels)) labels <- paste("PC", 1:2, sep="")
     if (length(labels)==1) labels <- paste(labels, 1:2, sep="")
     labels <- rep(labels, each = 2)
     text(result[label.ends,], 
          labels=labels[label.ends], 
-         pos=label.pos[label.ends], ...)
+         pos=label.pos[label.ends],
+         xpd = TRUE, ...)
   }
   invisible(result)
 
 }
 
-if(FALSE){
-# data(iris)
-cov <- cov(iris[,1:2])
-mu <- colMeans(iris[,1:2])
-
-radius <- sqrt(qchisq(0.68, 2))
-plot(iris[,1:2])
-car::ellipse(mu, cov, radius = radius)
-res <- ellipse.axes(cov, centre=mu, level = 0.68,
-                    labels = TRUE, 
-                    lwd = 2, col = "red", cex = 1.2)
-
-# try using arrows
-arrows(mu[1], mu[2], x1 = res[,1], y1 = res[,2], angle = 15, length = .15)
-
-plot(iris[,1:2])
-car::ellipse(mu, cov, radius = radius)
-res <- ellipse.axes(cov, centre=mu, level = 0.68,
-                    labels = TRUE, label.ends = 1:4)
-}
+# if(FALSE){
+# # data(iris)
+# cov <- cov(iris[,1:2])
+# mu <- colMeans(iris[,1:2])
+# 
+# radius <- sqrt(qchisq(0.68, 2))
+# plot(iris[,1:2])
+# car::ellipse(mu, cov, radius = radius)
+# res <- ellipse.axes(cov, centre=mu, level = 0.68,
+#                     labels = TRUE, 
+#                     lwd = 2, col = "red", cex = 1.2)
+# 
+# # try using arrows
+# arrows(mu[1], mu[2], x1 = res[,1], y1 = res[,2], angle = 15, length = .15)
+# 
+# plot(iris[,1:2])
+# car::ellipse(mu, cov, radius = radius)
+# res <- ellipse.axes(cov, centre=mu, level = 0.68,
+#                     labels = TRUE, label.ends = 1:4)
+# }